Move eventsource polyfill
