+++ packages/next/client/dev/event-source-polyfill.js
